[FIX] procurement_purchase_no_grouping: Fix no line grouping hack
Previous hack in v9, as there's a loop `for po_line in po.order_line` for
matching previous PO lines (this didn't happen in v8, as the match was done
through a restricted search), makes other parts that use the same loop to
fail. For example, PO totals computation.

This is another ugly hack, but it prevents the problem. The drawback is
that it needs to create/unlink a UoM record each time a procurement is
processed, as virtual records are not allowed here.